Add shadows
At each intersection from the camera rays, we generate a new ray towards
the light source (for each light source) and checks whether that ray
intersects scene objects. If so, the illumination from that light is
discarded, producing shadows. There are some caveats, such as the need
to displace the first intersection point (i.e. from the camera ray), to
avoid detecting the same intersection again when checking for occlusion.
The idea comes from [1].

[1]: https://github.com/ssloy/tinyraytracer/wiki/Part-1:-understandable-raytracing#step-6-shadows