Add specular light
This is based on the angle between the view ray and the
perfectly-reflected light ray. Note that we use {255, 255, 255} as the
base color for specular light instead of the sphere color. Otherwise,
the specular highlight and its surroundings would have the same color,
and we would not be able to see the highlight at all. Alternatively, we
could let the colors go beyond 255 and normalize the values later.
However, this would cause the rest of the image (i.e. besides the
specular highlight) to be too dark. The solution used here comes from
https://github.com/ssloy/tinyraytracer/blob/master/tinyraytracer.cpp#L124 .

Also note that, for now, all the spheres have the same reflecting
properties. I want to change that later, though.