ppm: use 'struct vec3' instead of the custom 'struct color'
Using vec3 is arguably better due to the number of already implemented
operators. This change simplifies the color manipulation code, although
it introduces two breaking changes: now colors are in the float domain
("ideally" 0.0 to 1.0) and the R, G, and B components must be accessed
through the x, y, and z fields. The second change is unfortunate, but
there is not much we can do about it... Regarding the first, I think it
comes down more as personal preference, and I quite like the float
range. We just have to convert the values back to [0, 255] when writing
the ppm to disk.