Fix issue with wrong sphere mapping at the poles
When a position maps to the H's pixel of a texture, we use fmod() and
end up taking the 0th (i.e. the first) pixel. This ends up producing a
discontinuation point at the north and south poles. Instead, let's get
rid of the fmod() and simply map the values to [0, H-1] and [0, W-1], as
expected.